refactor(core): remove misplaced sync services
