+++ packages/core/src/services/index.ts
@@ -1,5 +1,4 @@
 export * from "./clipboard";
 export * from "./randomness";
-export * from "./sync";
 export * from "./vault-session";
 export * from "./vault-snapshots";
