--- packages/core/src/services/sync/index.ts
@@ -1,3 +1,0 @@
-export * from "./sync-config.utils";
-export * from "./vault-sync-upload.service";
-export * from "./vault-sync-review.service";
