+++ packages/core/src/services/vault-snapshots/vault-snapshot.service.test.ts
@@ -256,27 +256,11 @@ describe("VaultSnapshotService", () => {
     expect(ctx.ports.crypto.decryptVaultSnapshotContent).not.toHaveBeenCalled();
   });
 
-  it("preflights unlocked vault persistence without saving a snapshot", async () => {
-    const ctx = createContext();
-
-    await ctx.service.requireUnlockedVaultCanBePersisted(
-      ctx.values.vaultId,
-      ctx.unlockedVault,
-      ctx.currentSnapshot.metadata.revision,
-    );
-
-    expect(ctx.ports.crypto.encryptVaultSnapshotContent).not.toHaveBeenCalled();
-    expect(ctx.ports.crypto.signVaultSnapshot).not.toHaveBeenCalled();
-    expect(
-      ctx.ports.vaultLocalRepository.saveVaultSnapshot,
-    ).not.toHaveBeenCalled();
-  });
-
-  it("fails preflight when the current local snapshot revision does not match the session source revision", async () => {
+  it("fails when the required current snapshot revision does not match the session source revision", async () => {
     const ctx = createContext();
 
     await expect(
-      ctx.service.requireUnlockedVaultCanBePersisted(
+      ctx.service.requireCurrentSnapshotForUnlockedVault(
         ctx.values.vaultId,
         ctx.unlockedVault,
         ctx.currentSnapshot.metadata.revision - 1,
