+++ packages/core/src/use-cases/sync/disable-sync.test.ts
@@ -68,14 +68,14 @@ describe("DisableSyncUseCase", () => {
     expect("syncConfig" in persistedUnlockedVault!.vault).toBe(false);
     expect(sourceSnapshotRevision).toBe(1);
     expect(
-      ctx.vaultSnapshot.requireUnlockedVaultCanBePersisted,
+      ctx.vaultSnapshot.requireCurrentSnapshotForUnlockedVault,
     ).toHaveBeenCalledWith(
       ctx.values.vaultId,
       persistedUnlockedVault,
       sourceSnapshotRevision,
     );
     expect(
-      vi.mocked(ctx.vaultSnapshot.requireUnlockedVaultCanBePersisted).mock
+      vi.mocked(ctx.vaultSnapshot.requireCurrentSnapshotForUnlockedVault).mock
         .invocationCallOrder[0],
     ).toBeLessThan(
       vi.mocked(ctx.ports.syncProvider.removeVaultSnapshots).mock
@@ -133,7 +133,7 @@ describe("DisableSyncUseCase", () => {
     });
 
     vi.mocked(
-      ctx.vaultSnapshot.requireUnlockedVaultCanBePersisted,
+      ctx.vaultSnapshot.requireCurrentSnapshotForUnlockedVault,
     ).mockRejectedValueOnce(error);
 
     await expect(
