+++ packages/core/src/use-cases/sync/require-vault-sync-config.ts
@@ -1,4 +1,3 @@
-import type { SyncConfig } from "../../domain/sync/sync-config.type";
 import type { Vault } from "../../domain/vault/vault";
 import { SyncNotConfiguredError } from "../../errors/sync.errors";
 
@@ -6,7 +5,7 @@ export function requireVaultSyncConfig(
   vaultId: string,
   operation: string,
   vault: Pick<Vault, "syncConfig">,
-): SyncConfig {
+) {
   const syncConfig = vault.syncConfig;
 
   if (syncConfig === undefined) {
