+++ packages/core/src/use-cases/sync/sync-upload.ts
@@ -1,25 +1,35 @@
-import { requireVaultSyncConfig } from "../../services/sync/sync-config.utils";
-import type { VaultSyncUploadService } from "../../services/sync/vault-sync-upload.service";
+import {
+  areRemoteVaultSnapshotDescriptorsEqual,
+  compareLocalAndRemoteSnapshotDescriptors,
+  toRemoteVaultSnapshotDescriptor,
+} from "../../domain/sync/vault-snapshot-version.utils";
+import {
+  RemoteVaultSnapshotAheadError,
+  RemoteVaultSnapshotChangedError,
+  SyncConflictDetectedError,
+} from "../../errors/sync.errors";
+import type { SyncProviderPort } from "../../ports/sync/sync-provider.port";
 import type { UnlockedVaultSessionService } from "../../services/vault-session/unlocked-vault-session.service";
 import type { VaultSnapshotService } from "../../services/vault-snapshots/vault-snapshot.service";
+import { requireVaultSyncConfig } from "./require-vault-sync-config";
 
 export type SyncUploadCommandParams = {
   readonly vaultId: string;
 };
 
 export class SyncUploadUseCase {
+  private readonly syncProvider: SyncProviderPort;
   private readonly unlockedVaultSession: UnlockedVaultSessionService;
   private readonly vaultSnapshot: VaultSnapshotService;
-  private readonly vaultSyncUpload: VaultSyncUploadService;
 
   constructor(
+    syncProvider: SyncProviderPort,
     unlockedVaultSession: UnlockedVaultSessionService,
     vaultSnapshot: VaultSnapshotService,
-    vaultSyncUpload: VaultSyncUploadService,
   ) {
+    this.syncProvider = syncProvider;
     this.unlockedVaultSession = unlockedVaultSession;
     this.vaultSnapshot = vaultSnapshot;
-    this.vaultSyncUpload = vaultSyncUpload;
   }
 
   async execute(params: SyncUploadCommandParams): Promise<void> {
@@ -37,12 +47,54 @@ export class SyncUploadUseCase {
     const localSnapshot = await this.vaultSnapshot.requireLocalVaultSnapshot(
       params.vaultId,
     );
+    const remoteSnapshotDescriptor =
+      await this.syncProvider.getLatestVaultSnapshotDescriptor(
+        syncConfig,
+        params.vaultId,
+      );
 
-    await this.vaultSyncUpload.uploadLocalSnapshotWhenSafe({
-      vaultId: params.vaultId,
-      syncConfig,
-      localVault: unlockedVault.vault,
-      localSnapshot,
-    });
+    if (remoteSnapshotDescriptor !== null) {
+      const localSnapshotDescriptor = toRemoteVaultSnapshotDescriptor(
+        localSnapshot.metadata.id,
+        unlockedVault.vault,
+        localSnapshot,
+      );
+      const relation = compareLocalAndRemoteSnapshotDescriptors(
+        localSnapshotDescriptor,
+        remoteSnapshotDescriptor,
+      );
+
+      if (
+        relation === "equal" &&
+        areRemoteVaultSnapshotDescriptorsEqual(
+          remoteSnapshotDescriptor,
+          localSnapshotDescriptor,
+        )
+      ) {
+        return;
+      }
+
+      if (relation === "equal" || relation === "remote_ahead") {
+        throw new RemoteVaultSnapshotAheadError(params.vaultId);
+      }
+
+      if (relation === "diverged") {
+        throw new SyncConflictDetectedError(params.vaultId);
+      }
+    }
+
+    try {
+      await this.syncProvider.uploadVaultSnapshot(
+        syncConfig,
+        localSnapshot,
+        remoteSnapshotDescriptor,
+      );
+    } catch (error) {
+      if (error instanceof RemoteVaultSnapshotChangedError) {
+        throw new SyncConflictDetectedError(params.vaultId);
+      }
+
+      throw error;
+    }
   }
 }
